kafka/requirements.txt
kafka-python==2.0.2
qdrant-client==1.7.0
fastembed==0.2.6
python-dotenv==1.0.0
